# kyros-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyros Backend",
    "description": "Wellness (Phase 1) + Clinic (Phase 2)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/wellness/tracked-items/": {
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "List Tracked Items",
        "operationId": "list_tracked_items_v1_wellness_tracked_items__get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "medication",
                    "water",
                    "meal",
                    "workout",
                    "vital_check",
                    "custom"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "paused",
                    "discontinued"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedItemRead"
                  },
                  "title": "Response List Tracked Items V1 Wellness Tracked Items  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Wellness"
        ],
        "summary": "Create Tracked Item",
        "operationId": "create_tracked_item_v1_wellness_tracked_items__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/MedicationItemCreate"
                  },
                  {
                    "$ref": "#/components/schemas/WaterItemCreate"
                  },
                  {
                    "$ref": "#/components/schemas/MealItemCreate"
                  },
                  {
                    "$ref": "#/components/schemas/WorkoutItemCreate"
                  },
                  {
                    "$ref": "#/components/schemas/VitalCheckItemCreate"
                  },
                  {
                    "$ref": "#/components/schemas/CustomItemCreate"
                  }
                ],
                "discriminator": {
                  "propertyName": "category",
                  "mapping": {
                    "medication": "#/components/schemas/MedicationItemCreate",
                    "water": "#/components/schemas/WaterItemCreate",
                    "meal": "#/components/schemas/MealItemCreate",
                    "workout": "#/components/schemas/WorkoutItemCreate",
                    "vital_check": "#/components/schemas/VitalCheckItemCreate",
                    "custom": "#/components/schemas/CustomItemCreate"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ]
      }
    },
    "/v1/wellness/tracked-items/{item_id}": {
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "Get Tracked Item",
        "operationId": "get_tracked_item_v1_wellness_tracked_items__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Wellness"
        ],
        "summary": "Update Tracked Item",
        "operationId": "update_tracked_item_v1_wellness_tracked_items__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackedItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Wellness"
        ],
        "summary": "Delete Tracked Item",
        "description": "Soft-delete: status \u2192 discontinued, all child reminders \u2192 inactive.",
        "operationId": "delete_tracked_item_v1_wellness_tracked_items__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/wellness/tracked-items/{item_id}/reminders": {
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "List Item Reminders",
        "operationId": "list_item_reminders_v1_wellness_tracked_items__item_id__reminders_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReminderRead"
                  },
                  "title": "Response List Item Reminders V1 Wellness Tracked Items  Item Id  Reminders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Wellness"
        ],
        "summary": "Create Item Reminder",
        "operationId": "create_item_reminder_v1_wellness_tracked_items__item_id__reminders_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{user_id}/edit": {
      "post": {
        "summary": "User Edit",
        "operationId": "user_edit_admin_users__user_id__edit_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_user_edit_admin_users__user_id__edit_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/items": {
      "get": {
        "summary": "Items List",
        "operationId": "items_list_admin_items_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Category"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Source"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "category": {
            "type": "string",
            "const": "custom",
            "title": "Category"
          },
          "metadata": {
            "$ref": "#/components/schemas/CustomMeta"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "category",
          "metadata"
        ],
        "title": "CustomItemCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "category": {
            "type": "string",
            "const": "meal",
            "title": "Category"
          },
          "metadata": {
            "$ref": "#/components/schemas/MealMeta"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "category",
          "metadata"
        ],
        "title": "MealItemCreate"
      },
...
      "MedicationItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "category": {
            "type": "string",
            "const": "medication",
            "title": "Category"
          },
          "metadata": {
            "$ref": "#/components/schemas/MedicationMeta"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "category",
          "metadata"
        ],
        "title": "MedicationItemCreate"
      },
...
      "ReminderCreate": {
        "properties": {
          "schedule": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RecurringSchedule"
              },
              {
                "$ref": "#/components/schemas/IntervalSchedule"
              }
            ],
            "title": "Schedule",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "interval": "#/components/schemas/IntervalSchedule",
                "recurring": "#/components/schemas/RecurringSchedule"
              }
            }
          },
          "message_template": {
            "type": "string",
            "title": "Message Template"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "schedule",
          "message_template"
        ],
        "title": "ReminderCreate"
      },
      "ReminderRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tracked_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tracked Item Id"
          },
          "schedule": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RecurringSchedule"
              },
              {
                "$ref": "#/components/schemas/IntervalSchedule"
              }
            ],
            "title": "Schedule",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "interval": "#/components/schemas/IntervalSchedule",
                "recurring": "#/components/schemas/RecurringSchedule"
              }
            }
          },
          "message_template": {
            "type": "string",
            "title": "Message Template"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tracked_item_id",
          "schedule",
          "message_template",
          "channels",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "ReminderRead"
      },
...
      "TrackedItemRead": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "medication",
              "water",
              "meal",
              "workout",
              "vital_check",
              "custom"
            ],
            "title": "Category"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "discontinued"
            ],
            "title": "Status"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "kyros",
              "ai_extracted"
            ],
            "title": "Source"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "reminders": {
            "items": {
              "$ref": "#/components/schemas/ReminderRead"
            },
            "type": "array",
            "title": "Reminders",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "category",
          "name",
          "id",
          "user_id",
          "status",
          "start_date",
          "end_date",
          "source",
          "source_ref",
          "created_at",
          "updated_at"
        ],
        "title": "TrackedItemRead"
      },
      "TrackedItemUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "item_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Metadata"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "paused",
                  "discontinued"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "title": "TrackedItemUpdate"
      },
...
      "VitalCheckItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "category": {
            "type": "string",
            "const": "vital_check",
            "title": "Category"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "category"
        ],
        "title": "VitalCheckItemCreate"
      },
      "WaterItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "category": {
            "type": "string",
            "const": "water",
            "title": "Category"
          },
          "metadata": {
            "$ref": "#/components/schemas/WaterMeta"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "category",
          "metadata"
        ],
        "title": "WaterItemCreate"
      },
...
      "WorkoutItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "category": {
            "type": "string",
            "const": "workout",
            "title": "Category"
          },
          "metadata": {
            "$ref": "#/components/schemas/WorkoutMeta"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "category",
          "metadata"
        ],
        "title": "WorkoutItemCreate"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}